0.2.0 - Minor (#52)
* :book: Included initial TODOs within code. Starting dev for v0.2.1.


* :hammer: Included experimental abstracted visualization workflow. Removed debug statement from epd_container.cpp.


* :fire: :hammer: Removed completed TODOs for Precision Level 3 Visualize workflow. Modified for common infer_action function call for specified workflow.


* :hammer: Rectified curLabel display in P3 Visualize Workflow.


* :hammer: Transferred experimental visualize function from EasyPerceptionDeployment to EPDContainer class for better unit testing coverage.


* :fire: Removed deprecated infer_visualize. Code verbosity reduced drastically.


* :hammer: Renamed infer_action for P3 U1 Visualize workflow to just infer.


* :hammer: Surfaced usecase_config activateUseCase function call to level of rclcpp::Node callback, EasyPerceptionDeployment. Ready for optimization in P3 U1 Visualize and Action Workflow.


* :hammer: Combined infer_action and infer_visualize for P2 U0 Visualize workflow. Modified EPDContainer visualize function to accept both P2 and P3 model detection outputs.


* :fire: :hammer: Removed deprecated infer_visualize-concerned functions within P2 workflow.


* :fire: Removed P2-specific activateUseCase, count and matchColor function calls. Generified usecase_config functions to adapt to both P2 and P3 model detection outputs. Verified with successfull local build.


* :hammer: Changed all instantiation of bboxes from data type float to int to silence intrinsic warnings for better debugging.


* :hammer: Corrected assignment of output_obj instead of result in P2 image callback. Added CHANGELOG.rst.


* :hammer: Attempted fix for failing CI build.


* :hammer: Adhered to ROS2 REP-2004 linting standards.


* :book: Updated copyright year to current.


* :hammer: Expanding EPD::activateUseCase for optimization. Yet to verify.


* :hammer: Removed all instances of inefficient .json file reading within usecase_config.hpp. Code verbosity reduced as well.


* :hammer: Replace infer_visualize call with infer_action for Tracking Visualize workflow. Verified with successful build and manual testing.n


* :hammer: Renamed infer_action in P2 workflows to infer.


* :fire: :hammer: Removed deprecated infer_visualize from Tracking workflow. Renamed infer_action to generic infer for same Tracking workflow.


* :hammer: Verified replaced of infer_visualize with infer_action within Localization workflow.


* :hammer: Renamed infer_action to infer for Localization workflow. Codestructuring completed.


* :hammer: Adhered codebase to REP-2004 linting standards.


* :hammer: Updated default session_config.json and input_image_topic.json.


* :book: :hammer: Rectified CHANGELOG.rst for 0.2.0 - Minor release.

+++ easy_perception_deployment/include/epd_utils_lib/message_utils.hpp
@@ -1,5 +1,5 @@
-// Copyright 2020 Advanced Remanufacturing and Technology Centre
-// Copyright 2020 ROS-Industrial Consortium Asia Pacific Team
+// Copyright 2022 Advanced Remanufacturing and Technology Centre
+// Copyright 2022 ROS-Industrial Consortium Asia Pacific Team
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -38,7 +38,7 @@ class EPDObjectDetection
 {
 public:
   /*! \brief A vector of bounding boxes with xmin, ymin, xmax, ymax.*/
-  std::vector<std::array<float, 4>> bboxes;
+  std::vector<std::array<int, 4>> bboxes;
   /*! \brief A vector of indices that indicate the numerical identities of
   corresponding bounding boxes of the same index.
   */
